Moved dims_utils.py → dimensions.py.
