+++ qutip/dimensions.py
